@@ -14,8 +14,9 @@
 from __future__ import annotations
 
 import warnings
+from collections.abc import Callable, Iterable
 from numbers import Number
-from typing import Any, Callable, Dict, Iterable, List, Optional, Set, Tuple, Type, TypeVar, Union
+from typing import Any, Optional, TypeVar, Union
 
 import numpy as np
 import oqpy
@@ -141,9 +142,9 @@ class Circuit:
 
         """
         self._moments: Moments = Moments()
-        self._result_types: Dict[ResultType] = {}
-        self._qubit_observable_mapping: Dict[Union[int, Circuit._ALL_QUBITS], Observable] = {}
-        self._qubit_observable_target_mapping: Dict[int, Tuple[int]] = {}
+        self._result_types: dict[ResultType] = {}
+        self._qubit_observable_mapping: dict[Union[int, Circuit._ALL_QUBITS], Observable] = {}
+        self._qubit_observable_target_mapping: dict[int, tuple[int]] = {}
         self._qubit_observable_set = set()
         self._parameters = set()
         self._observables_simultaneously_measurable = True
@@ -158,21 +159,21 @@ class Circuit:
         return self._moments.depth
 
     @property
-    def instructions(self) -> List[Instruction]:
+    def instructions(self) -> list[Instruction]:
         """Iterable[Instruction]: Get an `iterable` of instructions in the circuit."""
         return list(self._moments.values())
 
     @property
-    def result_types(self) -> List[ResultType]:
-        """List[ResultType]: Get a list of requested result types in the circuit."""
+    def result_types(self) -> list[ResultType]:
+        """list[ResultType]: Get a list of requested result types in the circuit."""
         return list(self._result_types.keys())
 
     @property
-    def basis_rotation_instructions(self) -> List[Instruction]:
+    def basis_rotation_instructions(self) -> list[Instruction]:
         """Gets a list of basis rotation instructions.
 
         Returns:
-            List[Instruction]: Get a list of basis rotation instructions in the circuit.
+            list[Instruction]: Get a list of basis rotation instructions in the circuit.
             These basis rotation instructions are added if result types are requested for
             an observable other than Pauli-Z.
 
@@ -200,8 +201,8 @@ class Circuit:
 
     @staticmethod
     def _observable_to_instruction(
-        observable: Observable, target_list: List[int]
-    ) -> List[Instruction]:
+        observable: Observable, target_list: list[int]
+    ) -> list[Instruction]:
         return [Instruction(gate, target_list) for gate in observable.basis_rotation_gates]
 
     @property
@@ -224,12 +225,12 @@ class Circuit:
         return QubitSet(self._moments.qubits.union(self._qubit_observable_set))
 
     @property
-    def parameters(self) -> Set[FreeParameter]:
+    def parameters(self) -> set[FreeParameter]:
         """
         Gets a set of the parameters in the Circuit.
 
         Returns:
-            Set[FreeParameter]: The `FreeParameters` in the Circuit.
+            set[FreeParameter]: The `FreeParameters` in the Circuit.
         """
         return self._parameters
 
@@ -237,7 +238,7 @@ class Circuit:
         self,
         result_type: ResultType,
         target: QubitSetInput = None,
-        target_mapping: Dict[QubitInput, QubitInput] = None,
+        target_mapping: dict[QubitInput, QubitInput] = None,
     ) -> Circuit:
         """
         Add a requested result type to `self`, returns `self` for chaining ability.
@@ -247,7 +248,7 @@ class Circuit:
             target (QubitSetInput): Target qubits for the
                 `result_type`.
                 Default = `None`.
-            target_mapping (Dict[QubitInput, QubitInput]): A dictionary of
+            target_mapping (dict[QubitInput, QubitInput]): A dictionary of
                 qubit mappings to apply to the `result_type.target`. Key is the qubit in
                 `result_type.target` and the value is what the key will be changed to.
                 Default = `None`.
@@ -377,7 +378,7 @@ class Circuit:
     @staticmethod
     def _tensor_product_index_dict(
         observable: TensorProduct, observable_target: QubitSet
-    ) -> Dict[int, Tuple[Observable, Tuple[int, ...]]]:
+    ) -> dict[int, tuple[Observable, tuple[int, ...]]]:
         obj_dict = {}
         i = 0
         factors = list(observable.factors)
@@ -401,7 +402,7 @@ class Circuit:
         self,
         instruction: Instruction,
         target: QubitSetInput = None,
-        target_mapping: Dict[QubitInput, QubitInput] = None,
+        target_mapping: dict[QubitInput, QubitInput] = None,
     ) -> Circuit:
         """
         Add an instruction to `self`, returns `self` for chaining ability.
@@ -412,7 +413,7 @@ class Circuit:
                 `instruction`. If a single qubit gate, an instruction is created for every index
                 in `target`.
                 Default = `None`.
-            target_mapping (Dict[QubitInput, QubitInput]): A dictionary of
+            target_mapping (dict[QubitInput, QubitInput]): A dictionary of
                 qubit mappings to apply to the `instruction.target`. Key is the qubit in
                 `instruction.target` and the value is what the key will be changed to.
                 Default = `None`.
@@ -493,7 +494,7 @@ class Circuit:
         self,
         circuit: Circuit,
         target: QubitSetInput = None,
-        target_mapping: Dict[QubitInput, QubitInput] = None,
+        target_mapping: dict[QubitInput, QubitInput] = None,
     ) -> Circuit:
         """
         Add a `circuit` to self, returns self for chaining ability.
@@ -504,7 +505,7 @@ class Circuit:
                 supplied circuit. This is a macro over `target_mapping`; `target` is converted to
                 a `target_mapping` by zipping together a sorted `circuit.qubits` and `target`.
                 Default = `None`.
-            target_mapping (Dict[QubitInput, QubitInput]): A dictionary of
+            target_mapping (dict[QubitInput, QubitInput]): A dictionary of
                 qubit mappings to apply to the qubits of `circuit.instructions`. Key is the qubit
                 to map, and the value is what to change it to. Default = `None`.
 
@@ -569,7 +570,7 @@ class Circuit:
         self,
         verbatim_circuit: Circuit,
         target: QubitSetInput = None,
-        target_mapping: Dict[QubitInput, QubitInput] = None,
+        target_mapping: dict[QubitInput, QubitInput] = None,
     ) -> Circuit:
         """
         Add a verbatim `circuit` to self, that is, ensures that `circuit` is not modified in any way
@@ -581,7 +582,7 @@ class Circuit:
                 supplied circuit. This is a macro over `target_mapping`; `target` is converted to
                 a `target_mapping` by zipping together a sorted `circuit.qubits` and `target`.
                 Default = `None`.
-            target_mapping (Dict[QubitInput, QubitInput]): A dictionary of
+            target_mapping (dict[QubitInput, QubitInput]): A dictionary of
                 qubit mappings to apply to the qubits of `circuit.instructions`. Key is the qubit
                 to map, and the value is what to change it to. Default = `None`.
 
@@ -637,8 +638,8 @@ class Circuit:
 
     def apply_gate_noise(
         self,
-        noise: Union[Type[Noise], Iterable[Type[Noise]]],
-        target_gates: Optional[Union[Type[Gate], Iterable[Type[Gate]]]] = None,
+        noise: Union[type[Noise], Iterable[type[Noise]]],
+        target_gates: Optional[Union[type[Gate], Iterable[type[Gate]]]] = None,
         target_unitary: np.ndarray = None,
         target_qubits: Optional[QubitSetInput] = None,
     ) -> Circuit:
@@ -664,9 +665,9 @@ class Circuit:
         only applied to gates with the same qubit_count in target_qubits.
 
         Args:
-            noise (Union[Type[Noise], Iterable[Type[Noise]]]): Noise channel(s) to be applied
+            noise (Union[type[Noise], Iterable[type[Noise]]]): Noise channel(s) to be applied
                 to the circuit.
-            target_gates (Optional[Union[Type[Gate], Iterable[Type[Gate]]]]): Gate class or
+            target_gates (Optional[Union[type[Gate], Iterable[type[Gate]]]]): Gate class or
                 List of Gate classes which `noise` is applied to. Default=None.
             target_unitary (ndarray): matrix of the target unitary gates. Default=None.
             target_qubits (Optional[QubitSetInput]): Index or indices of qubit(s).
@@ -781,7 +782,7 @@ class Circuit:
 
     def apply_initialization_noise(
         self,
-        noise: Union[Type[Noise], Iterable[Type[Noise]]],
+        noise: Union[type[Noise], Iterable[type[Noise]]],
         target_qubits: Optional[QubitSetInput] = None,
     ) -> Circuit:
         """Apply `noise` at the beginning of the circuit for every qubit (default) or
@@ -793,7 +794,7 @@ class Circuit:
         to `noise.qubit_count`.
 
         Args:
-            noise (Union[Type[Noise], Iterable[Type[Noise]]]): Noise channel(s) to be applied
+            noise (Union[type[Noise], Iterable[type[Noise]]]): Noise channel(s) to be applied
                 to the circuit.
             target_qubits (Optional[QubitSetInput]): Index or indices of qubit(s).
                 Default=None.
@@ -848,13 +849,13 @@ class Circuit:
 
         return apply_noise_to_moments(self, noise, target_qubits, "initialization")
 
-    def make_bound_circuit(self, param_values: Dict[str, Number], strict: bool = False) -> Circuit:
+    def make_bound_circuit(self, param_values: dict[str, Number], strict: bool = False) -> Circuit:
         """
         Binds FreeParameters based upon their name and values passed in. If parameters
         share the same name, all the parameters of that name will be set to the mapped value.
 
         Args:
-            param_values (Dict[str, Number]):  A mapping of FreeParameter names
+            param_values (dict[str, Number]):  A mapping of FreeParameter names
                 to a value to assign to them.
             strict (bool): If True, raises a ValueError if none of the FreeParameters
                 in param_values appear in the circuit. False by default."
@@ -867,12 +868,12 @@ class Circuit:
             self._validate_parameters(param_values)
         return self._use_parameter_value(param_values)
 
-    def _validate_parameters(self, parameter_values: Dict[str, Number]) -> None:
+    def _validate_parameters(self, parameter_values: dict[str, Number]) -> None:
         """
         This runs a check to see that the parameters are in the Circuit.
 
         Args:
-            parameter_values (Dict[str, Number]):  A mapping of FreeParameter names
+            parameter_values (dict[str, Number]):  A mapping of FreeParameter names
                 to a value to assign to them.
 
         Raises:
@@ -886,12 +887,12 @@ class Circuit:
             if param not in parameter_strings:
                 raise ValueError(f"No parameter in the circuit named: {param}")
 
-    def _use_parameter_value(self, param_values: Dict[str, Number]) -> Circuit:
+    def _use_parameter_value(self, param_values: dict[str, Number]) -> Circuit:
         """
         Creates a Circuit that uses the parameter values passed in.
 
         Args:
-            param_values (Dict[str, Number]): A mapping of FreeParameter names
+            param_values (dict[str, Number]): A mapping of FreeParameter names
                 to a value to assign to them.
 
         Returns:
@@ -932,7 +933,7 @@ class Circuit:
 
     def apply_readout_noise(
         self,
-        noise: Union[Type[Noise], Iterable[Type[Noise]]],
+        noise: Union[type[Noise], Iterable[type[Noise]]],
         target_qubits: Optional[QubitSetInput] = None,
     ) -> Circuit:
         """Apply `noise` right before measurement in every qubit (default) or target_qubits`.
@@ -943,7 +944,7 @@ class Circuit:
         to `noise.qubit_count`.
 
         Args:
-            noise (Union[Type[Noise], Iterable[Type[Noise]]]): Noise channel(s) to be applied
+            noise (Union[type[Noise], Iterable[type[Noise]]]): Noise channel(s) to be applied
                 to the circuit.
             target_qubits (Optional[QubitSetInput]): Index or indices of qubit(s).
                 Default=None.
@@ -1082,12 +1083,12 @@ class Circuit:
             circ.add_result_type(result_type)
         return circ
 
-    def diagram(self, circuit_diagram_class: Type = AsciiCircuitDiagram) -> str:
+    def diagram(self, circuit_diagram_class: type = AsciiCircuitDiagram) -> str:
         """
         Get a diagram for the current circuit.
 
         Args:
-            circuit_diagram_class (Type): A `CircuitDiagram` class that builds the
+            circuit_diagram_class (type): A `CircuitDiagram` class that builds the
                 diagram for this circuit. Default = `AsciiCircuitDiagram`.
 
         Returns:
@@ -1099,7 +1100,7 @@ class Circuit:
         self,
         ir_type: IRType = IRType.JAQCD,
         serialization_properties: SerializationProperties = None,
-        gate_definitions: Optional[Dict[Tuple[Gate, QubitSet], PulseSequence]] = None,
+        gate_definitions: Optional[dict[tuple[Gate, QubitSet], PulseSequence]] = None,
     ) -> Union[OpenQasmProgram, JaqcdProgram]:
         """
         Converts the circuit into the canonical intermediate representation.
@@ -1111,7 +1112,7 @@ class Circuit:
             serialization_properties (SerializationProperties): The serialization properties to use
                 while serializing the object to the IR representation. The serialization properties
                 supplied must correspond to the supplied `ir_type`. Defaults to None.
-            gate_definitions (Optional[Dict[Tuple[Gate, QubitSet], PulseSequence]]): The
+            gate_definitions (Optional[dict[tuple[Gate, QubitSet], PulseSequence]]): The
                 calibration data for the device. default: None.
 
         Returns:
@@ -1141,14 +1142,14 @@ class Circuit:
 
     @staticmethod
     def from_ir(
-        source: Union[str, OpenQasmProgram], inputs: Optional[Dict[str, io_type]] = None
+        source: Union[str, OpenQasmProgram], inputs: Optional[dict[str, io_type]] = None
     ) -> Circuit:
         """
         Converts an OpenQASM program to a Braket Circuit object.
 
         Args:
             source (Union[str, OpenQasmProgram]): OpenQASM string.
-            inputs (Optional[Dict[str, io_type]]): Inputs to the circuit.
+            inputs (Optional[dict[str, io_type]]): Inputs to the circuit.
 
         Returns:
             Circuit: Braket Circuit implementing the OpenQASM program.
@@ -1183,7 +1184,7 @@ class Circuit:
     def _to_openqasm(
         self,
         serialization_properties: OpenQASMSerializationProperties,
-        gate_definitions: Optional[Dict[Tuple[Gate, QubitSet], PulseSequence]],
+        gate_definitions: Optional[dict[tuple[Gate, QubitSet], PulseSequence]],
     ) -> OpenQasmProgram:
         ir_instructions = self._create_openqasm_header(serialization_properties, gate_definitions)
         openqasm_ir_type = IRType.OPENQASM
@@ -1220,8 +1221,8 @@ class Circuit:
     def _create_openqasm_header(
         self,
         serialization_properties: OpenQASMSerializationProperties,
-        gate_definitions: Optional[Dict[Tuple[Gate, QubitSet], PulseSequence]],
-    ) -> List[str]:
+        gate_definitions: Optional[dict[tuple[Gate, QubitSet], PulseSequence]],
+    ) -> list[str]:
         ir_instructions = ["OPENQASM 3.0;"]
         for parameter in self.parameters:
             ir_instructions.append(f"input float {parameter};")
@@ -1244,9 +1245,9 @@ class Circuit:
 
     def _validate_gate_calbrations_uniqueness(
         self,
-        gate_definitions: Dict[Tuple[Gate, QubitSet], PulseSequence],
-        frames: Dict[Frame],
-        waveforms: Dict[Waveform],
+        gate_definitions: dict[tuple[Gate, QubitSet], PulseSequence],
+        frames: dict[Frame],
+        waveforms: dict[ArbitraryWaveform],
     ) -> None:
         for key, calibration in gate_definitions.items():
             for frame in calibration._frames.values():
@@ -1257,7 +1258,7 @@ class Circuit:
                 waveforms[waveform.id] = waveform
 
     def _generate_frame_wf_defcal_declarations(
-        self, gate_definitions: Optional[Dict[Tuple[Gate, QubitSet], PulseSequence]]
+        self, gate_definitions: Optional[dict[tuple[Gate, QubitSet], PulseSequence]]
     ) -> Optional[str]:
         program = oqpy.Program(None, simplify_constants=False)
 
@@ -1303,8 +1304,8 @@ class Circuit:
         return None
 
     def _get_frames_waveforms_from_instrs(
-        self, gate_definitions: Optional[Dict[Tuple[Gate, QubitSet], PulseSequence]]
-    ) -> Tuple[Dict[Frame], Dict[Waveform]]:
+        self, gate_definitions: Optional[dict[tuple[Gate, QubitSet], PulseSequence]]
+    ) -> tuple[dict[Frame], dict[ArbitraryWaveform]]:
         from braket.circuits.gates import PulseGate
 
         frames = {}
@@ -1327,9 +1328,9 @@ class Circuit:
 
     def _add_fixed_argument_calibrations(
         self,
-        gate_definitions: Dict[Tuple[Gate, QubitSet], PulseSequence],
+        gate_definitions: dict[tuple[Gate, QubitSet], PulseSequence],
         instruction: Instruction,
-    ) -> Dict[Tuple[Gate, QubitSet], PulseSequence]:
+    ) -> dict[tuple[Gate, QubitSet], PulseSequence]:
         """Adds calibrations with arguments set to the instruction parameter values
 
         Given the collection of parameters in instruction.operator, this function looks for matching
@@ -1342,12 +1343,12 @@ class Circuit:
         If N=0, we ignore it as it will not be removed by _generate_frame_wf_defcal_declarations.
 
         Args:
-            gate_definitions (Dict[Tuple[Gate, QubitSet], PulseSequence]): a dictionary of
+            gate_definitions (dict[tuple[Gate, QubitSet], PulseSequence]): a dictionary of
                 calibrations
             instruction (Instruction): a Circuit instruction
 
         Returns:
-            Dict[Tuple[Gate, QubitSet], PulseSequence]: additional calibrations
+            dict[tuple[Gate, QubitSet], PulseSequence]: additional calibrations
 
         Raises:
             NotImplementedError: in two cases: (i) if the instruction contains unbound parameters
