@@ -14,7 +14,8 @@
 from __future__ import annotations
 
 import warnings
-from typing import TYPE_CHECKING, Any, Iterable, List, Optional, Tuple, Type, Union
+from collections.abc import Iterable
+from typing import TYPE_CHECKING, Any, Optional, Union
 
 import numpy as np
 
@@ -42,27 +43,27 @@ def no_noise_applied_warning(noise_applied: bool) -> None:
         )
 
 
-def wrap_with_list(an_item: Any) -> List[Any]:
+def wrap_with_list(an_item: Any) -> list[Any]:
     """Helper function to make the input parameter a list.
     Args:
         an_item (Any): The item to wrap.
 
     Returns:
-        List[Any]: The item wrapped in a list.
+        list[Any]: The item wrapped in a list.
     """
     if an_item is not None and not isinstance(an_item, list):
         an_item = [an_item]
     return an_item
 
 
-def check_noise_target_gates(noise: Noise, target_gates: Iterable[Type[Gate]]) -> None:
+def check_noise_target_gates(noise: Noise, target_gates: Iterable[type[Gate]]) -> None:
     """Helper function to check
     1. whether all the elements in target_gates are a Gate type;
     2. if `noise` is multi-qubit noise and `target_gates` contain gates
     with the number of qubits is the same as `noise.qubit_count`.
     Args:
         noise (Noise): A Noise class object to be applied to the circuit.
-        target_gates (Iterable[Type[Gate]]): Gate class or
+        target_gates (Iterable[type[Gate]]): Gate class or
             List of Gate classes which `noise` is applied to.
     """
     if not all(isinstance(g, type) and issubclass(g, Gate) for g in target_gates):
@@ -126,7 +127,7 @@ def check_noise_target_qubits(
 
 
 def apply_noise_to_moments(
-    circuit: Circuit, noise: Iterable[Type[Noise]], target_qubits: QubitSet, position: str
+    circuit: Circuit, noise: Iterable[type[Noise]], target_qubits: QubitSet, position: str
 ) -> Circuit:
     """
     Apply initialization/readout noise to the circuit.
@@ -138,7 +139,7 @@ def apply_noise_to_moments(
 
     Args:
         circuit (Circuit): A ciruit where `noise` is applied to.
-        noise (Iterable[Type[Noise]]): Noise channel(s) to be applied
+        noise (Iterable[type[Noise]]): Noise channel(s) to be applied
             to the circuit.
         target_qubits (QubitSet): Index or indices of qubits. `noise` is applied to.
         position (str): The position to add the noise to. May be 'initialization' or
@@ -181,18 +182,18 @@ def apply_noise_to_moments(
 
 
 def _apply_noise_to_gates_helper(
-    noise: Iterable[Type[Noise]],
+    noise: Iterable[type[Noise]],
     target_qubits: QubitSet,
     instruction: Instruction,
     noise_index: int,
     intersection: QubitSet,
     noise_applied: bool,
     new_noise_instruction: Iterable,
-) -> Tuple[Iterable[Instruction], int, bool]:
+) -> tuple[Iterable[Instruction], int, bool]:
     """Helper function to work out the noise instructions to be attached to a gate.
 
     Args:
-        noise (Iterable[Type[Noise]]): Noise channel(s) to be applied
+        noise (Iterable[type[Noise]]): Noise channel(s) to be applied
             to the circuit.
         target_qubits (QubitSet): Index or indices of qubits which `noise` is applied to.
         instruction (Instruction): Instruction of the gate which `noise` is applied to.
@@ -204,7 +205,7 @@ def _apply_noise_to_gates_helper(
             to the circuit.
 
     Returns:
-        Tuple[Iterable[Instruction], int, bool]: A tuple of three values:
+        tuple[Iterable[Instruction], int, bool]: A tuple of three values:
         new_noise_instruction: A list of noise intructions
         noise_index: The number of noise channels applied to the gate
         noise_applied: Whether noise is applied or not
@@ -234,8 +235,8 @@ def _apply_noise_to_gates_helper(
 
 def apply_noise_to_gates(
     circuit: Circuit,
-    noise: Iterable[Type[Noise]],
-    target_gates: Union[Iterable[Type[Gate]], np.ndarray],
+    noise: Iterable[type[Noise]],
+    target_gates: Union[Iterable[type[Gate]], np.ndarray],
     target_qubits: QubitSet,
 ) -> Circuit:
     """Apply noise after target gates in target qubits.
@@ -247,9 +248,9 @@ def apply_noise_to_gates(
 
     Args:
         circuit (Circuit): A ciruit where `noise` is applied to.
-        noise (Iterable[Type[Noise]]): Noise channel(s) to be applied
+        noise (Iterable[type[Noise]]): Noise channel(s) to be applied
             to the circuit.
-        target_gates (Union[Iterable[Type[Gate]], ndarray]): List of gates, or a unitary matrix
+        target_gates (Union[Iterable[type[Gate]], ndarray]): List of gates, or a unitary matrix
             which `noise` is applied to.
         target_qubits (QubitSet): Index or indices of qubits which `noise` is applied to.
 
