@@ -14,8 +14,9 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
+from collections.abc import Iterable
 from enum import Enum
-from typing import Any, Iterable, Set, Type, Union
+from typing import Any, Union
 
 
 class CriteriaKey(str, Enum):
@@ -54,14 +55,14 @@ class Criteria(ABC):
         raise NotImplementedError
 
     @abstractmethod
-    def get_keys(self, key_type: CriteriaKey) -> Union[CriteriaKeyResult, Set[Any]]:
+    def get_keys(self, key_type: CriteriaKey) -> Union[CriteriaKeyResult, set[Any]]:
         """Returns a set of key for a given key type.
 
         Args:
             key_type (CriteriaKey): The criteria key type.
 
         Returns:
-            Union[CriteriaKeyResult, Set[Any]]: Returns a set of keys for a key type. The
+            Union[CriteriaKeyResult, set[Any]]: Returns a set of keys for a key type. The
             actual returned keys will depend on the CriteriaKey. If the provided key type
             is not relevant the returned list will be empty. If the provided key type is
             relevant for all possible inputs, the string CriteriaKeyResult.ALL will be returned.
@@ -105,10 +106,10 @@ class Criteria(ABC):
         raise NotImplementedError
 
     @classmethod
-    def register_criteria(cls, criteria: Type[Criteria]) -> None:
+    def register_criteria(cls, criteria: type[Criteria]) -> None:
         """Register a criteria implementation by adding it into the Criteria class.
 
         Args:
-            criteria (Type[Criteria]): Criteria class to register.
+            criteria (type[Criteria]): Criteria class to register.
         """
         setattr(cls, criteria.__name__, criteria)
