@@ -13,7 +13,7 @@
 
 from __future__ import annotations
 
-from typing import Any, Dict, List, Type, Union
+from typing import Any, Union
 
 from braket.circuits.free_parameter import FreeParameter
 from braket.circuits.observable import Observable
@@ -34,10 +34,10 @@ class ResultType:
     the metadata that defines what a requested result type is and what it does.
     """
 
-    def __init__(self, ascii_symbols: List[str]):
+    def __init__(self, ascii_symbols: list[str]):
         """
         Args:
-            ascii_symbols (List[str]): ASCII string symbols for the result type. This is used when
+            ascii_symbols (list[str]): ASCII string symbols for the result type. This is used when
                 printing a diagram of circuits.
 
         Raises:
@@ -50,8 +50,8 @@ class ResultType:
         self._ascii_symbols = ascii_symbols
 
     @property
-    def ascii_symbols(self) -> List[str]:
-        """List[str]: Returns the ascii symbols for the requested result type."""
+    def ascii_symbols(self) -> list[str]:
+        """list[str]: Returns the ascii symbols for the requested result type."""
         return self._ascii_symbols
 
     @property
@@ -118,7 +118,7 @@ class ResultType:
         raise NotImplementedError("to_openqasm has not been implemented yet.")
 
     def copy(
-        self, target_mapping: Dict[QubitInput, QubitInput] = None, target: QubitSetInput = None
+        self, target_mapping: dict[QubitInput, QubitInput] = None, target: QubitSetInput = None
     ) -> ResultType:
         """
         Return a shallow copy of the result type.
@@ -128,7 +128,7 @@ class ResultType:
             qubits. This is useful apply an instruction to a circuit and change the target qubits.
 
         Args:
-            target_mapping (Dict[QubitInput, QubitInput]): A dictionary of
+            target_mapping (dict[QubitInput, QubitInput]): A dictionary of
                 qubit mappings to apply to the target. Key is the qubit in this `target` and the
                 value is what the key is changed to. Default = `None`.
             target (QubitSetInput): Target qubits for the new instruction.
@@ -162,11 +162,11 @@ class ResultType:
         return copy
 
     @classmethod
-    def register_result_type(cls, result_type: Type[ResultType]) -> None:
+    def register_result_type(cls, result_type: type[ResultType]) -> None:
         """Register a result type implementation by adding it into the `ResultType` class.
 
         Args:
-            result_type (Type[ResultType]): `ResultType` class to register.
+            result_type (type[ResultType]): `ResultType` class to register.
         """
         setattr(cls, result_type.__name__, result_type)
 
@@ -188,11 +188,11 @@ class ObservableResultType(ResultType):
     """
 
     def __init__(
-        self, ascii_symbols: List[str], observable: Observable, target: QubitSetInput = None
+        self, ascii_symbols: list[str], observable: Observable, target: QubitSetInput = None
     ):
         """
         Args:
-            ascii_symbols (List[str]): ASCII string symbols for the result type. This is used when
+            ascii_symbols (list[str]): ASCII string symbols for the result type. This is used when
                 printing a diagram of circuits.
             observable (Observable): the observable for the result type
             target (QubitSetInput): Target qubits that the
@@ -285,10 +285,10 @@ class ObservableParameterResultType(ObservableResultType):
 
     def __init__(
         self,
-        ascii_symbols: List[str],
+        ascii_symbols: list[str],
         observable: Observable,
         target: QubitSetInput = None,
-        parameters: List[Union[str, FreeParameter]] = None,
+        parameters: list[Union[str, FreeParameter]] = None,
     ):
         super().__init__(ascii_symbols, observable, target)
 
@@ -300,13 +300,13 @@ class ObservableParameterResultType(ObservableResultType):
 
         """
         Args:
-            ascii_symbols (List[str]): ASCII string symbols for the result type. This is used when
+            ascii_symbols (list[str]): ASCII string symbols for the result type. This is used when
                 printing a diagram of circuits.
             observable (Observable): the observable for the result type.
             target (QubitSetInput): Target qubits that the result type is requested for.
                 Default is `None`, which means the observable must only operate on 1
                 qubit and it will be applied to all qubits in parallel.
-            parameters (List[Union[str, FreeParameter]]): List of string inputs or
+            parameters (list[Union[str, FreeParameter]]): List of string inputs or
                 FreeParameter objects. These inputs will be used as parameters for
                 gradient calculation. Default: `all`.
 
@@ -318,7 +318,7 @@ class ObservableParameterResultType(ObservableResultType):
         """
 
     @property
-    def parameters(self) -> List[str]:
+    def parameters(self) -> list[str]:
         return self._parameters
 
     def __repr__(self) -> str:
