@@ -15,7 +15,7 @@ from __future__ import annotations
 
 import re
 from functools import reduce
-from typing import List, Union
+from typing import Union
 
 import braket.ir.jaqcd as ir
 from braket.circuits import circuit
@@ -178,19 +178,19 @@ class AdjointGradient(ObservableParameterResultType):
     def __init__(
         self,
         observable: Observable,
-        target: List[QubitSetInput] = None,
-        parameters: List[Union[str, FreeParameter]] = None,
+        target: list[QubitSetInput] = None,
+        parameters: list[Union[str, FreeParameter]] = None,
     ):
         """
         Args:
             observable (Observable): The expectation value of this observable is the function
                 against which parameters in the gradient are differentiated.
-            target (List[QubitSetInput]): Target qubits that the result type is requested for.
+            target (list[QubitSetInput]): Target qubits that the result type is requested for.
                 Each term in the target list should have the same number of qubits as the
                 corresponding term in the observable. Default is `None`, which means the
                 observable must operate only on 1 qubit and it is applied to all qubits
                 in parallel.
-            parameters (List[Union[str, FreeParameter]]): The free parameters in the circuit to
+            parameters (list[Union[str, FreeParameter]]): The free parameters in the circuit to
                 differentiate with respect to. Default: `all`.
 
         Raises:
@@ -240,20 +240,20 @@ class AdjointGradient(ObservableParameterResultType):
     @circuit.subroutine(register=True)
     def adjoint_gradient(
         observable: Observable,
-        target: List[QubitSetInput] = None,
-        parameters: List[Union[str, FreeParameter]] = None,
+        target: list[QubitSetInput] = None,
+        parameters: list[Union[str, FreeParameter]] = None,
     ) -> ResultType:
         """Registers this function into the circuit class.
 
         Args:
             observable (Observable): The expectation value of this observable is the function
                 against which parameters in the gradient are differentiated.
-            target (List[QubitSetInput]): Target qubits that the result type is requested for.
+            target (list[QubitSetInput]): Target qubits that the result type is requested for.
                 Each term in the target list should have the same number of qubits as the
                 corresponding term in the observable. Default is `None`, which means the
                 observable must operate only on 1 qubit and it is applied to all qubits
                 in parallel.
-            parameters (List[Union[str, FreeParameter]]): The free parameters in the circuit to
+            parameters (list[Union[str, FreeParameter]]): The free parameters in the circuit to
                 differentiate with respect to. Default: `all`.
 
         Returns:
@@ -279,10 +279,10 @@ class Amplitude(ResultType):
     This is available on simulators only when `shots=0`.
     """
 
-    def __init__(self, state: List[str]):
+    def __init__(self, state: list[str]):
         """
         Args:
-            state (List[str]): list of quantum states as strings with "0" and "1"
+            state (list[str]): list of quantum states as strings with "0" and "1"
 
         Raises:
             ValueError: If state is `None` or an empty list, or
@@ -293,7 +293,7 @@ class Amplitude(ResultType):
         """
         if (
             not state
-            or not isinstance(state, List)
+            or not isinstance(state, list)
             or not all(
                 isinstance(amplitude, str) and re.fullmatch("^[01]+$", amplitude)
                 for amplitude in state
@@ -306,7 +306,7 @@ class Amplitude(ResultType):
         self._state = state
 
     @property
-    def state(self) -> List[str]:
+    def state(self) -> list[str]:
         return self._state
 
     def _to_jaqcd(self) -> ir.Amplitude:
@@ -318,11 +318,11 @@ class Amplitude(ResultType):
 
     @staticmethod
     @circuit.subroutine(register=True)
-    def amplitude(state: List[str]) -> ResultType:
+    def amplitude(state: list[str]) -> ResultType:
         """Registers this function into the circuit class.
 
         Args:
-            state (List[str]): list of quantum states as strings with "0" and "1"
+            state (list[str]): list of quantum states as strings with "0" and "1"
 
         Returns:
             ResultType: state vector as a requested result type
