@@ -11,16 +11,14 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-from typing import List
-
 from braket.device_schema import error_mitigation
 
 
 class ErrorMitigation:
-    def serialize(self) -> List[error_mitigation.ErrorMitigationScheme]:
+    def serialize(self) -> list[error_mitigation.ErrorMitigationScheme]:
         """
         Returns:
-            List[ErrorMitigationScheme]: A list of service-readable error
+            list[ErrorMitigationScheme]: A list of service-readable error
             mitigation scheme descriptions.
         """
         raise NotImplementedError("serialize is not implemented.")
