@@ -10,8 +10,11 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
+
+from __future__ import annotations
+
 from pathlib import Path
-from typing import Any, Dict, Union
+from typing import Any
 
 from braket.jobs.environment_variables import get_checkpoint_dir, get_job_name, get_results_dir
 from braket.jobs.serialization import deserialize_values, serialize_values
@@ -19,7 +22,7 @@ from braket.jobs_data import PersistedJobData, PersistedJobDataFormat
 
 
 def save_job_checkpoint(
-    checkpoint_data: Dict[str, Any],
+    checkpoint_data: dict[str, Any],
     checkpoint_file_suffix: str = "",
     data_format: PersistedJobDataFormat = PersistedJobDataFormat.PLAINTEXT,
 ) -> None:
@@ -35,7 +38,7 @@ def save_job_checkpoint(
 
 
     Args:
-        checkpoint_data (Dict[str, Any]): Dict that specifies the checkpoint data to be persisted.
+        checkpoint_data (dict[str, Any]): Dict that specifies the checkpoint data to be persisted.
         checkpoint_file_suffix (str): str that specifies the file suffix to be used for
             the checkpoint filename. The resulting filename
             `f"{job_name}(_{checkpoint_file_suffix}).json"` is used to save the checkpoints.
@@ -62,27 +65,31 @@ def save_job_checkpoint(
         f.write(persisted_data.json())
 
 
-def load_job_checkpoint(job_name: str, checkpoint_file_suffix: str = "") -> Dict[str, Any]:
+def load_job_checkpoint(
+    job_name: str | None = None, checkpoint_file_suffix: str = ""
+) -> dict[str, Any]:
     """
-    Loads the hybrid job checkpoint data stored for the job named 'job_name', with the checkpoint
-    file that ends with the `checkpoint_file_suffix`. The `job_name` can refer to any hybrid job
-    whose checkpoint data you expect to be available in the file path specified by the
-    `CHECKPOINT_DIR` container environment variable.
+    Loads the job checkpoint data stored for the job named 'job_name', with the checkpoint
+    file that ends with the `checkpoint_file_suffix`. The `job_name` can refer to any job whose
+    checkpoint data you expect to be available in the file path specified by the `CHECKPOINT_DIR`
+    container environment variable. If not provided, this function will use the currently running
+    job's name.
 
     Note: This function for loading hybrid job checkpoints is only for use inside the job container
           as it writes data to directories and references env variables set in the containers.
 
 
     Args:
-        job_name (str): str that specifies the name of the hybrid job whose checkpoints
-            are to be loaded.
+        job_name (str | None): str that specifies the name of the job whose checkpoints
+            are to be loaded. Default: current job name.
+
         checkpoint_file_suffix (str): str specifying the file suffix that is used to
             locate the checkpoint file to load. The resulting file name
             `f"{job_name}(_{checkpoint_file_suffix}).json"` is used to locate the
             checkpoint file. Default: ""
 
     Returns:
-        Dict[str, Any]: Dict that contains the checkpoint data persisted in the checkpoint file.
+        dict[str, Any]: Dict that contains the checkpoint data persisted in the checkpoint file.
 
     Raises:
         FileNotFoundError: If the file `f"{job_name}(_{checkpoint_file_suffix})"` could not be found
@@ -90,6 +97,7 @@ def load_job_checkpoint(job_name: str, checkpoint_file_suffix: str = "") -> Dict
         ValueError: If the data stored in the checkpoint file can't be deserialized (possibly due to
             corruption).
     """
+    job_name = job_name or get_job_name()
     checkpoint_directory = get_checkpoint_dir()
     checkpoint_file_path = (
         f"{checkpoint_directory}/{job_name}_{checkpoint_file_suffix}.json"
@@ -104,7 +112,7 @@ def load_job_checkpoint(job_name: str, checkpoint_file_suffix: str = "") -> Dict
         return deserialized_data
 
 
-def _load_persisted_data(filename: Union[str, Path] = None) -> PersistedJobData:
+def _load_persisted_data(filename: str | Path = None) -> PersistedJobData:
     filename = filename or Path(get_results_dir()) / "results.json"
     try:
         with open(filename, mode="r") as f:
@@ -116,17 +124,17 @@ def _load_persisted_data(filename: Union[str, Path] = None) -> PersistedJobData:
         )
 
 
-def load_job_result(filename: Union[str, Path] = None) -> Dict[str, Any]:
+def load_job_result(filename: str | Path = None) -> dict[str, Any]:
     """
     Loads job result of currently running job.
 
     Args:
-        filename (Union[str, Path]): Location of job results. Default `results.json` in job
+        filename (str | Path): Location of job results. Default `results.json` in job
             results directory in a job instance or in working directory locally. This file
             must be in the format used by `save_job_result`.
 
     Returns:
-        Dict[str, Any]: Job result data of current job
+        dict[str, Any]: Job result data of current job
     """
     persisted_data = _load_persisted_data(filename)
     deserialized_data = deserialize_values(persisted_data.dataDictionary, persisted_data.dataFormat)
@@ -134,7 +142,7 @@ def load_job_result(filename: Union[str, Path] = None) -> Dict[str, Any]:
 
 
 def save_job_result(
-    result_data: Union[Dict[str, Any], Any],
+    result_data: dict[str, Any] | Any,
     data_format: PersistedJobDataFormat = None,
 ) -> None:
     """
@@ -147,7 +155,7 @@ def save_job_result(
 
 
     Args:
-        result_data (Union[Dict[str, Any], Any]): Dict that specifies the result data to be
+        result_data (dict[str, Any] | Any): Dict that specifies the result data to be
             persisted. If result data is not a dict, then it will be wrapped as
             `{"result": result_data}`.
         data_format (PersistedJobDataFormat): The data format used to serialize the
@@ -178,7 +186,7 @@ def save_job_result(
         current_persisted_data.dataDictionary,
         current_persisted_data.dataFormat,
     )
-    updated_results = {**current_results, **result_data}
+    updated_results = current_results | result_data
 
     with open(Path(get_results_dir()) / "results.json", "w") as f:
         serialized_data = serialize_values(updated_results or {}, data_format)
