@@ -19,12 +19,13 @@ import sys
 import tarfile
 import tempfile
 import time
+import warnings
+from collections.abc import Callable
 from dataclasses import asdict
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import Any
 
 from braket.aws.aws_session import AwsSession
-from braket.jobs import Framework, image_uris
 from braket.jobs.config import (
     CheckpointConfig,
     DeviceConfig,
@@ -33,6 +34,7 @@ from braket.jobs.config import (
     S3DataSourceConfig,
     StoppingCondition,
 )
+from braket.jobs.image_uris import Framework, retrieve_image
 
 
 def prepare_quantum_job(
@@ -43,8 +45,8 @@ def prepare_quantum_job(
     job_name: str = None,
     code_location: str = None,
     role_arn: str = None,
-    hyperparameters: Dict[str, Any] = None,
-    input_data: Union[str, Dict, S3DataSourceConfig] = None,
+    hyperparameters: dict[str, Any] = None,
+    input_data: str | dict | S3DataSourceConfig = None,
     instance_config: InstanceConfig = None,
     distribution: str = None,
     stopping_condition: StoppingCondition = None,
@@ -52,13 +54,17 @@ def prepare_quantum_job(
     copy_checkpoints_from_job: str = None,
     checkpoint_config: CheckpointConfig = None,
     aws_session: AwsSession = None,
-    tags: Dict[str, str] = None,
-) -> Dict:
+    tags: dict[str, str] = None,
+) -> dict:
     """Creates a hybrid job by invoking the Braket CreateJob API.
 
     Args:
-        device (str): ARN for the AWS device which is primarily
-            accessed for the execution of this hybrid job.
+        device (str): Device ARN of the QPU device that receives priority quantum
+            task queueing once the hybrid job begins running. Each QPU has a separate hybrid jobs
+            queue so that only one hybrid job is running at a time. The device string is accessible
+            in the hybrid job instance as the environment variable "AMZN_BRAKET_DEVICE_ARN".
+            When using embedded simulators, you may provide the device argument as string of the
+            form: "local:<provider>/<simulator_name>".
 
         source_module (str): Path (absolute, relative or an S3 URI) to a python module to be
             tarred and uploaded. If `source_module` is an S3 URI, it must point to a
@@ -86,12 +92,12 @@ def prepare_quantum_job(
         role_arn (str): A str providing the IAM role ARN used to execute the
             script. Default: IAM role returned by AwsSession's `get_default_jobs_role()`.
 
-        hyperparameters (Dict[str, Any]): Hyperparameters accessible to the hybrid job.
+        hyperparameters (dict[str, Any]): Hyperparameters accessible to the hybrid job.
             The hyperparameters are made accessible as a Dict[str, str] to the hybrid job.
             For convenience, this accepts other types for keys and values, but `str()`
             is called to convert them before being passed on. Default: None.
 
-        input_data (Union[str, Dict, S3DataSourceConfig]): Information about the training
+        input_data (str | dict | S3DataSourceConfig): Information about the training
             data. Dictionary maps channel names to local paths or S3 URIs. Contents found
             at any local paths will be uploaded to S3 at
             f's3://{default_bucket_name}/jobs/{job_name}/data/{channel_name}. If a local
@@ -99,9 +105,9 @@ def prepare_quantum_job(
             channel name "input".
             Default: {}.
 
-        instance_config (InstanceConfig): Configuration of the instances to be used
-            to execute the hybrid job. Default: InstanceConfig(instanceType='ml.m5.large',
-            instanceCount=1, volumeSizeInGB=30, volumeKmsKey=None).
+        instance_config (InstanceConfig): Configuration of the instance(s) for running the
+            classical code for the hybrid job. Defaults to
+            `InstanceConfig(instanceType='ml.m5.large', instanceCount=1, volumeSizeInGB=30)`.
 
         distribution (str): A str that specifies how the hybrid job should be distributed. If set to
             "data_parallel", the hyperparameters for the hybrid job will be set to use data
@@ -130,11 +136,11 @@ def prepare_quantum_job(
         aws_session (AwsSession): AwsSession for connecting to AWS Services.
             Default: AwsSession()
 
-        tags (Dict[str, str]): Dict specifying the key-value pairs for tagging this hybrid job.
+        tags (dict[str, str]): Dict specifying the key-value pairs for tagging this hybrid job.
             Default: {}.
 
     Returns:
-        Dict: Hybrid job tracking the execution on Amazon Braket.
+        dict: Hybrid job tracking the execution on Amazon Braket.
 
     Raises:
         ValueError: Raises ValueError if the parameters are not valid.
@@ -149,7 +155,7 @@ def prepare_quantum_job(
     _validate_params(param_datatype_map)
     aws_session = aws_session or AwsSession()
     device_config = DeviceConfig(device)
-    job_name = job_name or _generate_default_job_name(image_uri)
+    job_name = job_name or _generate_default_job_name(image_uri=image_uri)
     role_arn = role_arn or os.getenv("BRAKET_JOBS_ROLE_ARN", aws_session.get_default_jobs_role())
     hyperparameters = hyperparameters or {}
     hyperparameters = {str(key): str(value) for key, value in hyperparameters.items()}
@@ -181,7 +187,7 @@ def prepare_quantum_job(
             "compressionType": "GZIP",
         }
     }
-    image_uri = image_uri or image_uris.retrieve_image(Framework.BASE, aws_session.region)
+    image_uri = image_uri or retrieve_image(Framework.BASE, aws_session.region)
     algorithm_specification["containerImage"] = {"uri": image_uri}
     if not output_data_config.s3Path:
         output_data_config.s3Path = AwsSession.construct_s3_uri(
@@ -226,24 +232,37 @@ def prepare_quantum_job(
     return create_job_kwargs
 
 
-def _generate_default_job_name(image_uri: Optional[str]) -> str:
+def _generate_default_job_name(image_uri: str | None = None, func: Callable | None = None) -> str:
     """
-    Generate default hybrid job name using the image uri and a timestamp
+    Generate default job name using the image uri and entrypoint function.
+
     Args:
-        image_uri (Optional[str]): URI for the image container.
+        image_uri (str | None): URI for the image container.
+        func (Callable | None): The entry point function.
 
     Returns:
         str: Hybrid job name.
     """
-    if not image_uri:
-        job_type = "-default"
-    else:
-        job_type_match = re.search("/amazon-braket-(.*)-jobs:", image_uri) or re.search(
-            "/amazon-braket-([^:/]*)", image_uri
-        )
-        job_type = f"-{job_type_match.groups()[0]}" if job_type_match else ""
+    max_length = 50
+    timestamp = str(int(time.time() * 1000))
 
-    return f"braket-job{job_type}-{time.time() * 1000:.0f}"
+    if func:
+        name = func.__name__.replace("_", "-")
+        if len(name) + len(timestamp) > max_length:
+            name = name[: max_length - len(timestamp) - 1]
+            warnings.warn(
+                f"Job name exceeded {max_length} characters. Truncating name to {name}-{timestamp}."
+            )
+    else:
+        if not image_uri:
+            name = "braket-job-default"
+        else:
+            job_type_match = re.search("/amazon-braket-(.*)-jobs:", image_uri) or re.search(
+                "/amazon-braket-([^:/]*)", image_uri
+            )
+            container = f"-{job_type_match.groups()[0]}" if job_type_match else ""
+            name = f"braket-job{container}"
+    return f"{name}-{timestamp}"
 
 
 def _process_s3_source_module(
@@ -337,7 +356,7 @@ def _tar_and_upload_to_code_location(
         aws_session.upload_to_s3(f"{temp_dir}/source.tar.gz", f"{code_location}/source.tar.gz")
 
 
-def _validate_params(dict_arr: Dict[str, Tuple[any, any]]) -> None:
+def _validate_params(dict_arr: dict[str, tuple[any, any]]) -> None:
     """
     Validate that config parameters are of the right type.
 
@@ -356,19 +375,19 @@ def _validate_params(dict_arr: Dict[str, Tuple[any, any]]) -> None:
 
 
 def _process_input_data(
-    input_data: Union[str, Dict, S3DataSourceConfig], job_name: str, aws_session: AwsSession
-) -> List[Dict[str, Any]]:
+    input_data: str | dict | S3DataSourceConfig, job_name: str, aws_session: AwsSession
+) -> list[dict[str, Any]]:
     """
     Convert input data into a list of dicts compatible with the Braket API.
     Args:
-        input_data (Union[str, Dict, S3DataSourceConfig]): Either a channel definition or a
+        input_data (str | dict | S3DataSourceConfig): Either a channel definition or a
             dictionary mapping channel names to channel definitions, where a channel definition
             can be an S3DataSourceConfig or a str corresponding to a local prefix or S3 prefix.
         job_name (str): Hybrid job name.
         aws_session (AwsSession): AwsSession for possibly uploading local data.
 
     Returns:
-        List[Dict[str, Any]]: A list of channel configs.
+        list[dict[str, Any]]: A list of channel configs.
     """
     if not isinstance(input_data, dict):
         input_data = {"input": input_data}
@@ -405,17 +424,17 @@ def _process_channel(
         return S3DataSourceConfig(s3_prefix)
 
 
-def _convert_input_to_config(input_data: Dict[str, S3DataSourceConfig]) -> List[Dict[str, Any]]:
+def _convert_input_to_config(input_data: dict[str, S3DataSourceConfig]) -> list[dict[str, Any]]:
     """
     Convert a dictionary mapping channel names to S3DataSourceConfigs into a list of channel
     configs compatible with the Braket API.
 
     Args:
-        input_data (Dict[str, S3DataSourceConfig]): A dictionary mapping channel names to
+        input_data (dict[str, S3DataSourceConfig]): A dictionary mapping channel names to
             S3DataSourceConfig objects.
 
     Returns:
-        List[Dict[str, Any]]: A list of channel configs.
+        list[dict[str, Any]]: A list of channel configs.
     """
     return [
         {
@@ -426,5 +445,5 @@ def _convert_input_to_config(input_data: Dict[str, S3DataSourceConfig]) -> List[
     ]
 
 
-def _exclude_nones_factory(items: List[Tuple]) -> Dict:
+def _exclude_nones_factory(items: list[tuple]) -> dict:
     return {k: v for k, v in items if v is not None}
