@@ -11,7 +11,7 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-from typing import Any, Dict, Optional
+from typing import Any, Optional
 
 from oqpy import PortVar
 from oqpy.base import OQPyExpression
@@ -23,12 +23,12 @@ class Port:
     a device. See https://openqasm.com/language/openpulse.html#ports for more details.
     """
 
-    def __init__(self, port_id: str, dt: float, properties: Optional[Dict[str, Any]] = None):
+    def __init__(self, port_id: str, dt: float, properties: Optional[dict[str, Any]] = None):
         """
         Args:
             port_id (str): str identifying a unique port on the device.
             dt (float): The smallest time step that may be used on the control hardware.
-            properties (Optional[Dict[str, Any]]): Dict containing properties of
+            properties (Optional[dict[str, Any]]): Dict containing properties of
                 this port. Defaults to None.
         """
         self._port_id = port_id
